Improved UiToIdeContentProposalProvider
