+++ org.eclipse.xtext.ui/src/org/eclipse/xtext/ui/editor/contentassist/AbstractContentProposalProvider.java
@@ -32,7 +32,7 @@ import com.google.inject.Inject;
  * @author Heiko Behrens
  * @author Jan K&ouml;hnlein
  */
-public abstract class AbstractContentProposalProvider implements IContentProposalProvider, ICompletionProposalFactory {
+public abstract class AbstractContentProposalProvider extends AbstractCompletionProposalFactory implements IContentProposalProvider {
 
 	public class DefaultContentAssistProcessorSwitch extends XtextSwitch<Boolean> implements IFollowElementAcceptor {
 
@@ -74,9 +74,6 @@ public abstract class AbstractContentProposalProvider implements IContentProposa
 	}
 
 	@Inject
-	private IContentProposalPriorities priorities;
-
-	@Inject
 	@ContentProposalLabelProvider
 	private ILabelProvider labelProvider;
 
@@ -85,9 +82,6 @@ public abstract class AbstractContentProposalProvider implements IContentProposa
 
 	@Inject
 	private IValueConverterService valueConverter;
-
-	@Inject
-	private IProposalConflictHelper conflictHelper;
 
 	public static class NullSafeCompletionProposalAcceptor extends ICompletionProposalAcceptor.Delegate {
 
@@ -122,71 +116,6 @@ public abstract class AbstractContentProposalProvider implements IContentProposa
 
 	public IFollowElementAcceptor createSelector(ContentAssistContext context, ICompletionProposalAcceptor acceptor) {
 		return new DefaultContentAssistProcessorSwitch(context, acceptor);
-	}
-
-	public ICompletionProposal createCompletionProposal(String proposal, String displayString, Image image,
-			ContentAssistContext contentAssistContext) {
-		return createCompletionProposal(proposal, new StyledString(displayString), image, getPriorityHelper()
-				.getDefaultPriority(), contentAssistContext.getPrefix(), contentAssistContext);
-	}
-
-	protected ICompletionProposal createCompletionProposal(String proposal, ContentAssistContext contentAssistContext) {
-		return createCompletionProposal(proposal, null, null, getPriorityHelper().getDefaultPriority(),
-				contentAssistContext.getPrefix(), contentAssistContext);
-	}
-
-	@Override
-	public ICompletionProposal createCompletionProposal(String proposal, StyledString displayString, Image image,
-			ContentAssistContext contentAssistContext) {
-		return createCompletionProposal(proposal, displayString, image, getPriorityHelper().getDefaultPriority(),
-				contentAssistContext.getPrefix(), contentAssistContext);
-	}
-
-	/**
-	 * @see #isValidProposal(String, String, ContentAssistContext)
-	 * @see #doCreateProposal(String, StyledString, Image, int, ContentAssistContext)
-	 */
-	protected ICompletionProposal createCompletionProposal(String proposal, StyledString displayString, Image image,
-			int priority, String prefix, ContentAssistContext context) {
-		if (isValidProposal(proposal, prefix, context)) {
-			return doCreateProposal(proposal, displayString, image, priority, context);
-		}
-		return null;
-	}
-
-	protected boolean isValidProposal(String proposal, String prefix, ContentAssistContext context) {
-		if (proposal == null)
-			return false;
-		if (!context.getMatcher().isCandidateMatchingPrefix(proposal, prefix))
-			return false;
-		if (conflictHelper.existsConflict(proposal, context))
-			return false;
-		return true;
-	}
-
-	protected ConfigurableCompletionProposal doCreateProposal(String proposal, StyledString displayString, Image image,
-			int priority, ContentAssistContext context) {
-		int replacementOffset = context.getReplaceRegion().getOffset();
-		int replacementLength = context.getReplaceRegion().getLength();
-		ConfigurableCompletionProposal result = doCreateProposal(proposal, displayString, image, replacementOffset,
-				replacementLength);
-		result.setPriority(priority);
-		result.setMatcher(context.getMatcher());
-		result.setReplaceContextLength(getReplacementContextLength(context));
-		return result;
-	}
-
-	/**
-	 * @since 2.12
-	 */
-	protected int getReplacementContextLength(ContentAssistContext context) {
-		return context.getReplaceContextLength();
-	}
-
-	protected ConfigurableCompletionProposal doCreateProposal(String proposal, StyledString displayString, Image image,
-			int replacementOffset, int replacementLength) {
-		return new ConfigurableCompletionProposal(proposal, replacementOffset, replacementLength, proposal.length(),
-				image, displayString, null, null);
 	}
 
 	protected StyledString getKeywordDisplayString(Keyword keyword) {
@@ -251,21 +180,7 @@ public abstract class AbstractContentProposalProvider implements IContentProposa
 		return labelProvider.getImage(eObject);
 	}
 
-	public void setConflictHelper(IProposalConflictHelper conflictHelper) {
-		this.conflictHelper = conflictHelper;
-	}
-
-	public IProposalConflictHelper getConflictHelper() {
-		return conflictHelper;
-	}
-
-	public void setPriorityHelper(IContentProposalPriorities priorities) {
-		this.priorities = priorities;
-	}
-
-	public IContentProposalPriorities getPriorityHelper() {
-		return priorities;
-	}
+	
 
 	public IQualifiedNameConverter getQualifiedNameConverter() {
 		return qualifiedNameConverter;
